Google Analytics
1) Added settings from django_integration_settings (had to update that package)
2) Modified base.html to include necessary tags in an 'extra_scripts' block
3) Modified some settings in base.py to make variables available in context for this
4) Modified requirements to use '-e git+https' for integration settings.

For the last piece, installation via the .zip method did not include the template or template
tags from this submodule, despite updating the setup.py and manifest.in accordingly. I have
no idea why, and in the end I was wasting a great deal of time trying to troubleshoot this
when it's really not worth the effort. Switched to '-e git+https' unless someone else
wants to take a crack at it.