Fix Tests and Form
Tests were failing becuase slug was a required field that was now editable,
so it was on the "Submit a new resource" form.

Users shouldn't have to manually specify this, so I added blank=True and
excluded it in SubmitResourceForm's Meta class. This way it's still
auto-populated when saved, but will be editable from admin after the fact.