Fix case study save method
Needed to move copying of date fields into ContentType instead, as doing it
within the inheriting class required one to call Super twice, once to save and
trigger the publication field to change, and again if that is now "published"
to save the copy over to the date_created field. This worked just fine, but in
the tests it caused a duplicate key error for some reason. Moving this into
ContentType.save() instead, with a simple conditional to only trigger for case
studies, solved all of these issues.