Partners Tab
Fixes #147.

Added partners tab to browse view. RSS Feed from SCPD is accessed in the
BrowseView.get_context_data() method; feedparser.parse() is called to retrieve
and process the XML document and pass as a new object to the template.

A new include template was created to display the data within the tab. It
simply outputs the title and sets up a block_content for an intro paragraph,
then loops through all members in the feed and renders them in panels. Logo
display is restricted according to the same level criteria as in the SCPD
list view.

Logos and titles all link to the full member profile on the SCPD as some
have not yet provided any links that we could use instead.