Cosmetic Tweaks + Fix Tests
Fixed broken tests. Needed to use __str__(self) method of the topic
object. For some reason it kept saying it had no attribute "name", though
if I tried to print that attribute within the test, it worked just fine.

Also added the topic name to the title line of the first panel. This was
needed to facilitate the caching test, as the only thing we can vary is
that name. It's slightly redundant, but looks ok.