Fix Broken Report Links
django-linkcheck constructs links to the objects containing the broken link
fields solely by invoking "object.object.get_absolute_url". With no way to
override this short of issuing a PR to the package to fix this narrow method,
I added that method to each relevant object.

In each, since ContentType is a foreignkey relation, I just invoke
self.ct.get_admin_url() to point it right at the admin panel to edit the
link. Saves a step of having to go to the content piece then enter the edit.