Content Type Landing Page - Part 1
Fixed Issue #295.

Added tab toggles to content type pages.

Moved Resources list into Resources tab.

Created layout and various blocks for Summary tab. Intro for topic
pages is pulled from an object attribute "introduction", but in the
case of the ContentType model, the actual content type is just a
CharField attribute from a list of choices, with no metadata object
to tack on an intro field.

As such, it was simplest to simply create the intros as block_content.

Added data to the get_context_data() method in the view to display the
various "at a glance" data points requested.

TO-DO:

1) Fix graph divs. Currently, despite toggling, all appear in the source
upon loading and only visibility is affected. This causes my text text
to be displayed as if all 3 divs were there stacked on one another.

2) Add graphs and map. Probably need to add a visualization package.