Summary Map
Added google map to the Summary Map tab of the content type landing page.

Added custom JS to only render map if the page is loaded with #map-tab OR
when that tab is selected, and then hides the map if any other tab is selected,
otherwise you end up with a hidden map generating awkward whitespace.

That just leaves the Discipline graph display bug and Issue #295 should be done.