Graph Display Bug Fix
Not actually sure how I fixed this. Initial attempts at displaying both
graphs in the same tab in child divs failed, but I must have changed something
else and forgot to revert it? Whatever, it works now!

Added an option to the graphs to allow 20% on the left for the labels to avoid
excessive truncation and awkward whitespace to the right.