Fix Map Data
The map geocode data was the hardest to do. The API is very specific about
encoding and formatting. I couldn't just use the orgs context variable list,
it did end up requiring a whole new object that parses the queryset and
creates a new context variable containing the properly formatted (utf8) set
of data points.

The up-side is it required no modification of the map template, as I was
able to match the original data object 1-for-1.