Another Map Bug!
Turns out it's NOT required to have an organization associated with a
ContentType object! This resulted in an error with the new list comprehension
method of parsing the data, where an empty attribute "organizations" created
a "NoneType has no attribute encode" exception.

Similarly, if an organization does not have geocoding data, it kicked an
error "cannot convert string to float" when trying to convert None to a
decimal.

Had to use Q() to construct exclusion clauses into the query to omit these
corner cases.