Map Links
Added additional annotations to the map data to retrieve a count of
resources for that institution in the queryset for this type, and to
construct the URL to link to the filtered browse view preloaded with
this content type and the organization specified.

Updated map script to construct the necessary html tags and display
those in the infowindow in place of what we originally just the name.