Add fancy counters!
Revised the counter panel to display the stats using the fancy counters
from Unify instead of just plain text.

JS files only work if imported in base.html because of dependence on jquery,
for some reason that fails if imported in a template extending it instead.
I just wrapped them in a conditional to only load if it's a content type page,
it isn't loaded if it's any other type of page, and that seemed sufficient.