Chart Display Bug Quickfix
Rare bug encountered where if a user directly loaded the URL with #map-panel,
when clicking into the charts tab it displayed an error because it had
attempted to load the graphs with dimensions 0x0. Did not occur if the graphs
were loaded, then the user went to the map tab and back.

Added an explicit call to the chart initializing functions as onclick events
to catch this weird corner case.