Attempted Chart Bug Fix
Added a custom getChart() function through which we can pass an argument from
the html a-tags. This is then checked and used to set the display property
of the two tabs, allowing us to then explicitly call the chart drawing
function AFTER that property is switched, hopefully averting the problem
of the API calls being made before the div is visible.