Alter Chart Styling
I think I finally figured out the stupid ChartArea attribute! Those values are padding
around JUST the chart, and that's how much space is left for things like titles
and axis values and stuff.

The percentage is relative to the containing div, so for
the left and right sides that makes sense so it scales for mobile, but for top and
bottom, the height changes dramatically based on the number of rows! So a chart
with 40 has a HUGE margin, while the small 6-7 item charts in dev look just fine. That's
why editing those percentages still left us with the bug in production, because
our test environments don't have enough data.

Setting these margins as absolute pixels solves that. The font size is fixed, so setting
an absolute value like that is just fine...we always need the same space for the text.