Thumbnail Spacing Bug
Fixes #355.

Intermittent issue with spacing on some resources with images was finally
traced to be due to variations in height of the thumbnail divs. If the first
image in the series had the tallest div (due to the longest description under
the image), then the first div in the second row would be placed to the right
of it in order to appear as close to the top of the page as possible.

Setting the height of these divs to always be 400 pixels (sufficient to
contain the largest one we have that I can find...we shouldn't need more
space) resolves the issue since now the second row still tries going as
high on the page as possible, but that's always possible in the leftmost
column now.