Add Orgs to truncate_db Command
Discovered cause of "permission denied" error on dev instances. We are way
over the limit of 10,000 rows despite truncating the DB. There are many causes
including sessions and thumbnails, both of which have their own commands to
clear out the tables.

But we had over 8000 organizations being synced via ISS, which right off the
bat is 4/5 of our limit. Truncating that to 1000, which should fix the problem
but leave enough intact that we should still have enough data to test with.