Content Type Landing Page Bug Fixes
Fixed bug where links to tabs within the introduction block were not working
by auto-generating buttons instead.

Added function to use as onclick event for "View All" button which alters
the necessary properties of the relevant tab divs to manually mimic the
effects of actually clicking on the tab directly (aria-expanded boolean
switches, and class of tabs changes from None to active and vice versa).

Had to amend the context to provide access to the singular of the content
type name. For some reason we only had plurals everywhere.