Right Align Images
Also figured out why wrapping text and image in separate col-md-x divs
breaks the spacing. Need to also wrap the <hr> in a div or else it just
floats to the right of the item with width=0. Adding a div wrapper to
it makes it span the full width and add the missing padding.