Fix Metadata Display for Detail Pages
Initially just Fixes #393

Fixed that quickly, but then discovered this mysterious "institution" it
was attempting to display. Looks like someone created that but then this
metadata was called "Organizations" instead, but no one bothered to go back
and update it.

Added a conditional and loop to display the list of associated orgs for
each content type fragment that referenced this.